Specialized Room Tracker (S.R.T)_Octagon - BACKUP/requirements.txt
PySide6-Essentials
mysql-connector-python
`